Refactor(Icons): Rename icons/components
