+++ src/tokens/icons/assets/CloseAllFaults.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-function CloseAllFaults(props: React.SVGProps<SVGSVGElement>, svgRef?: React.Ref<SVGSVGElement>) {
+function SvgComponent(props: React.SVGProps<SVGSVGElement>, svgRef?: React.Ref<SVGSVGElement>) {
 	return (
 		<svg
 			{...props}
@@ -17,5 +17,5 @@ function CloseAllFaults(props: React.SVGProps<SVGSVGElement>, svgRef?: React.Ref
 	);
 }
 
-const ForwardRef = React.forwardRef(CloseAllFaults);
+const ForwardRef = React.forwardRef(SvgComponent);
 export default ForwardRef;
